SegmentMetadataQuery: Fix default interval handling. (#5489) (#5603)
* SegmentMetadataQuery: Fix default interval handling.

PR #4131 introduced a new copy builder for segmentMetadata that did
not retain the value of usingDefaultInterval. This led to it being
dropped and the default-interval handling not working as expected.
Instead of using the default 1 week history when intervals are not
provided, the segmentMetadata query would query _all_ segments,
incurring an unexpected performance hit.

This patch fixes the bug and adds a test for the copy builder.

* Intervals
(cherry picked from commit a35fe63141313816dadf42ac995f95c63217e2ff)

Change-Id: If55f8741b21075e92efd7160bf1d60cd7c19253c